+++ src/summary.h
@@ -23,27 +23,27 @@ public:
     // report[0] = starts, report[1] = ends, report[2] = lengths,
     // report[3] = ambigs, report[4] = homopolymers, report[5] = num_ns
     Rcpp::DataFrame summarizeFasta(vector<vector<int>> report,
-                                   vector<int> counts);
+                                   vector<float> counts);
 
-    Rcpp::DataFrame summarize(Rcpp::DataFrame, vector<int> counts,
+    Rcpp::DataFrame summarize(Rcpp::DataFrame, vector<float> counts,
                               vector<string> dfNames);
 private:
     int processors;
-    long long total, numUniques;
+    double total, numUniques;
 
     // dataframe column names -> map of values and counts
     // "starts" -> (position -> number of seqs with that start position)
     // "ends" -> (position -> number of seqs with that end position)
-    map<string, map<double, long long> > results;
+    map<string, map<double, double> > results;
 
     void createThreadsFasta(vector<vector<int>>& fasta,
-                                          vector<int>& counts);
+                                          vector<float>& counts);
 
     void createThreadsReport(vector<vector<double>>& report,
-                            vector<int>& counts, vector<string>&);
+                            vector<float>& counts, vector<string>&);
 
-    vector<double> getValues(map<double, long long>& positions);
-    vector<long long> getDefaults();
+    vector<double> getValues(map<double, double>& positions);
+    vector<double> getDefaults();
 };
 /******************************************************************************/
